requirements.txt
tkinter
opencv-python
numpy
pillow